+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/tasks/TickerTask.java
@@ -1,5 +1,6 @@
 package io.github.thebusybiscuit.slimefun4.implementation.tasks;
 
+import java.util.Collections;
 import java.util.HashSet;
 import java.util.Iterator;
 import java.util.Map;
@@ -10,17 +11,18 @@ import java.util.logging.Level;
 import javax.annotation.Nonnull;
 import javax.annotation.ParametersAreNonnullByDefault;
 
+import org.apache.commons.lang.Validate;
 import org.bukkit.Bukkit;
+import org.bukkit.Chunk;
 import org.bukkit.Location;
 import org.bukkit.Material;
-import org.bukkit.World;
 import org.bukkit.block.Block;
 import org.bukkit.scheduler.BukkitScheduler;
 
 import io.github.thebusybiscuit.cscorelib2.blocks.BlockPosition;
+import io.github.thebusybiscuit.cscorelib2.blocks.ChunkPosition;
 import io.github.thebusybiscuit.slimefun4.api.ErrorReport;
 import io.github.thebusybiscuit.slimefun4.implementation.SlimefunPlugin;
-import io.github.thebusybiscuit.slimefun4.utils.PatternUtils;
 import me.mrCookieSlime.CSCoreLibPlugin.Configuration.Config;
 import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SlimefunItem;
 import me.mrCookieSlime.Slimefun.Objects.handlers.BlockTicker;
@@ -39,17 +41,17 @@ import me.mrCookieSlime.Slimefun.api.Slimefun;
 public class TickerTask implements Runnable {
 
     /**
-     *  This Map holds all currently actively ticking locations.
+     * This Map holds all currently actively ticking locations.
      */
-    private final Map<String, Set<Location>> activeTickers = new ConcurrentHashMap<>();
+    private final Map<ChunkPosition, Set<Location>> tickingLocations = new ConcurrentHashMap<>();
 
     // These are "Queues" of blocks that need to be removed or moved
     private final Map<Location, Location> movingQueue = new ConcurrentHashMap<>();
     private final Map<Location, Boolean> deletionQueue = new ConcurrentHashMap<>();
 
     /**
-     *  This Map tracks how many bugs have occurred in a given Location .
-     *  If too many bugs happen, we delete that Location.
+     * This Map tracks how many bugs have occurred in a given Location .
+     * If too many bugs happen, we delete that Location.
      */
     private final Map<BlockPosition, Integer> bugs = new ConcurrentHashMap<>();
 
@@ -97,8 +99,8 @@ public class TickerTask implements Runnable {
             }
 
             if (!halted) {
-                for (Map.Entry<String, Set<Location>> entry : activeTickers.entrySet()) {
-                    tickChunk(tickers, entry.getKey(), entry.getValue());
+                for (Map.Entry<ChunkPosition, Set<Location>> entry : tickingLocations.entrySet()) {
+                    tickChunk(entry.getKey(), tickers, entry.getValue());
                 }
             }
 
@@ -123,21 +125,16 @@ public class TickerTask implements Runnable {
     }
 
     @ParametersAreNonnullByDefault
-    private void tickChunk(Set<BlockTicker> tickers, String chunk, Set<Location> locations) {
+    private void tickChunk(ChunkPosition chunk, Set<BlockTicker> tickers, Set<Location> locations) {
         try {
-            String[] components = PatternUtils.SEMICOLON.split(chunk);
-
-            World world = Bukkit.getWorld(components[0]);
-            int x = Integer.parseInt(components[components.length - 2]);
-            int z = Integer.parseInt(components[components.length - 1]);
-
-            if (world != null && world.isChunkLoaded(x, z)) {
+            // Only continue if the Chunk is actually loaded
+            if (chunk.isLoaded()) {
                 for (Location l : locations) {
                     tickLocation(tickers, l);
                 }
             }
         } catch (ArrayIndexOutOfBoundsException | NumberFormatException x) {
-            Slimefun.getLogger().log(Level.SEVERE, x, () -> "An Exception has occurred while trying to parse Chunk: " + chunk);
+            Slimefun.getLogger().log(Level.SEVERE, x, () -> "An Exception has occurred while trying to resolve Chunk: " + chunk);
         }
     }
 
@@ -235,14 +232,82 @@ public class TickerTask implements Runnable {
     }
 
     /**
-     * This method returns the {@link Map} of actively ticking locations according to
-     * their chunk id.
+     * This method returns a <strong>read-only</strong> {@link Map}
+     * representation of every {@link ChunkPosition} and its corresponding
+     * {@link Set} of ticking {@link Location Locations}.
      * 
-     * @return The {@link Map} of active tickers
+     * This does include any {@link Location} from an unloaded {@link Chunk} too!
+     * 
+     * @return A {@link Map} representation of all ticking {@link Location Locations}
      */
     @Nonnull
-    public Map<String, Set<Location>> getActiveTickers() {
-        return activeTickers;
+    public Map<ChunkPosition, Set<Location>> getLocations() {
+        return Collections.unmodifiableMap(tickingLocations);
+    }
+
+    /**
+     * This method returns a <strong>read-only</strong> {@link Set}
+     * of all ticking {@link Location Locations} in a given {@link Chunk}.
+     * The {@link Chunk} does not have to be loaded.
+     * If no {@link Location} is present, the returned {@link Set} will be empty.
+     * 
+     * @param chunk
+     *            The {@link Chunk}
+     * 
+     * @return A {@link Set} of all ticking {@link Location Locations}
+     */
+    @Nonnull
+    public Set<Location> getLocations(@Nonnull Chunk chunk) {
+        Validate.notNull(chunk, "The Chunk cannot be null!");
+
+        Set<Location> locations = tickingLocations.getOrDefault(new ChunkPosition(chunk), new HashSet<>());
+        return Collections.unmodifiableSet(locations);
+    }
+
+    /**
+     * This enables the ticker at the given {@link Location} and adds it to our "queue".
+     * 
+     * @param l
+     *            The {@link Location} to activate
+     */
+    public void enableTicker(@Nonnull Location l) {
+        Validate.notNull(l, "Location cannot be null!");
+
+        ChunkPosition chunk = new ChunkPosition(l.getWorld(), l.getBlockX() >> 4, l.getBlockZ() >> 4);
+        Set<Location> newValue = new HashSet<>();
+        Set<Location> oldValue = tickingLocations.putIfAbsent(chunk, newValue);
+
+        /**
+         * This is faster than doing computeIfAbsent(...)
+         * on a ConcurrentHashMap because it won't block the Thread for too long
+         */
+        if (oldValue != null) {
+            oldValue.add(l);
+        } else {
+            newValue.add(l);
+        }
+    }
+
+    /**
+     * This method disables the ticker at the given {@link Location} and removes it from our internal
+     * "queue".
+     * 
+     * @param l
+     *            The {@link Location} to remove
+     */
+    public void disableTicker(@Nonnull Location l) {
+        Validate.notNull(l, "Location cannot be null!");
+
+        ChunkPosition chunk = new ChunkPosition(l.getWorld(), l.getBlockX() >> 4, l.getBlockZ() >> 4);
+        Set<Location> locations = tickingLocations.get(chunk);
+
+        if (locations != null) {
+            locations.remove(l);
+
+            if (locations.isEmpty()) {
+                tickingLocations.remove(chunk);
+            }
+        }
     }
 
 }