+++ src/main/java/me/mrCookieSlime/Slimefun/api/BlockStorage.java
@@ -7,10 +7,8 @@ import java.nio.file.Files;
 import java.nio.file.StandardCopyOption;
 import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.Map;
 import java.util.Optional;
-import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.logging.Level;
 import java.util.logging.Logger;
@@ -69,10 +67,6 @@ public class BlockStorage {
 
     private static String serializeLocation(Location l) {
         return l.getWorld().getName() + ';' + l.getBlockX() + ';' + l.getBlockY() + ';' + l.getBlockZ();
-    }
-
-    private static String locationToChunkString(Location l) {
-        return l.getWorld().getName() + ";Chunk;" + (l.getBlockX() >> 4) + ';' + (l.getBlockZ() >> 4);
     }
 
     private static String serializeChunk(World world, int x, int z) {
@@ -177,7 +171,6 @@ public class BlockStorage {
         }
 
         try {
-            String chunkString = locationToChunkString(l);
             String json = cfg.getString(key);
             Config blockInfo = parseBlockInfo(l, json);
 
@@ -195,9 +188,7 @@ public class BlockStorage {
                 storage.put(l, blockInfo);
 
                 if (SlimefunPlugin.getRegistry().getTickerBlocks().contains(file.getName().replace(".sfb", ""))) {
-                    Map<String, Set<Location>> tickers = SlimefunPlugin.getTickerTask().getActiveTickers();
-                    Set<Location> locations = tickers.computeIfAbsent(chunkString, id -> new HashSet<>());
-                    locations.add(l);
+                    SlimefunPlugin.getTickerTask().enableTicker(l);
                 }
             }
         } catch (Exception x) {
@@ -495,11 +486,7 @@ public class BlockStorage {
         }
     }
 
-    public static void setBlockInfo(Block block, Config cfg, boolean updateTicker) {
-        setBlockInfo(block.getLocation(), cfg, updateTicker);
-    }
-
-    public static void setBlockInfo(Location l, Config cfg, boolean updateTicker) {
+    private static void setBlockInfo(Location l, Config cfg, boolean updateTicker) {
         BlockStorage storage = getStorage(l.getWorld());
 
         if (storage == null) {
@@ -590,17 +577,7 @@ public class BlockStorage {
                 universalInventory.save();
             }
 
-            String chunkString = locationToChunkString(l);
-            Map<String, Set<Location>> tickers = SlimefunPlugin.getTickerTask().getActiveTickers();
-            Set<Location> locations = tickers.get(chunkString);
-
-            if (locations != null) {
-                locations.remove(l);
-
-                if (locations.isEmpty()) {
-                    tickers.remove(chunkString);
-                }
-            }
+            SlimefunPlugin.getTickerTask().disableTicker(l);
         }
     }
 
@@ -638,23 +615,15 @@ public class BlockStorage {
         refreshCache(storage, from, previousData.getString("id"), null, true);
         storage.storage.remove(from);
 
-        String chunkString = locationToChunkString(from);
-        Map<String, Set<Location>> tickers = SlimefunPlugin.getTickerTask().getActiveTickers();
-        Set<Location> locations = tickers.get(chunkString);
-
-        if (locations != null) {
-            locations.remove(from);
-
-            if (locations.isEmpty()) {
-                tickers.remove(chunkString);
-            }
-        }
+        SlimefunPlugin.getTickerTask().disableTicker(from);
     }
 
     private static void refreshCache(BlockStorage storage, Location l, String key, String value, boolean updateTicker) {
         if (key == null) {
-            // This Block is no longer valid...
-            // Fixes #1577
+            /**
+             * This Block is no longer valid...
+             * Fixes #1577
+             */
             return;
         }
 
@@ -664,14 +633,8 @@ public class BlockStorage {
         if (updateTicker) {
             SlimefunItem item = SlimefunItem.getByID(key);
 
-            if (item != null && item.isTicking()) {
-                String chunkString = locationToChunkString(l);
-
-                if (value != null) {
-                    Map<String, Set<Location>> tickers = SlimefunPlugin.getTickerTask().getActiveTickers();
-                    Set<Location> locations = tickers.computeIfAbsent(chunkString, id -> new HashSet<>());
-                    locations.add(l);
-                }
+            if (item != null && item.isTicking() && value != null) {
+                SlimefunPlugin.getTickerTask().enableTicker(l);
             }
         }
     }